requirements.txt
tensorflow-gpu
keras 
numpy 
skimage 
matplotlib
